Move FileFlags class to FileFlags.(cpp|h) files
